visual_ai_studio/requirements_extra.txt
streamlit>=1.33
python-dotenv>=1.0
openai>=1.0.0
requests>=2.31
pillow>=10.0
